movie_recommendation/back_server/requirements.txt
absl-py==1.4.0
anyio==3.6.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.5.2
asttokens==2.0.8
astunparse==1.6.3
async-generator==1.10
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.11.2
bleach==6.0.0
cachetools==5.3.0
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.5
comm==0.1.2
contourpy==1.0.7
cryptography==40.0.2
cycler==0.11.0
debugpy==1.6.5
decorator==5.1.1
defusedxml==0.7.1
dj-rest-auth==4.0.0
Django==3.2.13
django-allauth==0.52.0
django-cors-headers==3.14.0
django-extensions==3.2.0
django-seed==0.2.2
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
entrypoints==0.4
exceptiongroup==1.1.1
executing==1.0.0
Faker==8.1.1
fastjsonschema==2.16.3
filelock==3.10.0
flatbuffers==23.3.3
fonttools==4.39.0
fqdn==1.5.1
gast==0.4.0
google-auth==2.16.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.3
h11==0.14.0
h5py==3.8.0
huggingface-hub==0.13.2
idna==3.3
image==1.5.33
importlib-metadata==6.0.0
importlib-resources==5.12.0
ipykernel==6.20.2
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==7.4.9
jupyter_core==5.1.3
jupyter_server==2.4.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
keras==2.11.0
kiwisolver==1.4.4
libclang==15.0.6.1
lightgbm==3.3.5
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==2.0.5
nbclassic==0.5.3
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
nltk==3.8.1
notebook==6.5.3
notebook_shim==0.2.2
numpy==1.22.4
oauthlib==3.2.0
opendata-kr==0.0.4
opt-einsum==3.3.0
outcome==1.2.0
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==2.6.2
prometheus-client==0.16.0
prompt-toolkit==3.0.30
protobuf==3.19.6
psutil==5.9.4
psycopg2==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.13.0
PyJWT==1.7.1
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-json-logger==2.0.7
python3-openid==3.2.0
pytz==2021.3
pywin32==305
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.0.0
qtconsole==5.4.0
QtPy==2.3.0
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
selenium==4.8.2
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
sqlparse==0.4.2
stack-data==0.5.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
terminado==0.17.1
text-unidecode==1.3
threadpoolctl==3.1.0
tinycss2==1.2.1
tokenizers==0.13.2
toposort==1.7
tornado==6.2
tqdm==4.65.0
traitlets==5.3.0
transformers==4.27.1
trio==0.22.0
trio-websocket==0.10.0
typing_extensions==4.5.0
uri-template==1.2.0
urllib3==1.26.8
wcwidth==0.2.5
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.2.3
widgetsnbextension==4.0.5
wrapt==1.15.0
wsproto==1.2.0
zipp==3.15.0
